requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
Werkzeug
psycopg2-binary